Add permission-based access control

+++ inverno-ticket/src/main/java/io/inverno/app/ticket/internal/rest/v1/PlanWebController.java
@@ -16,17 +16,20 @@
 package io.inverno.app.ticket.internal.rest.v1;
 
 import io.inverno.core.annotation.Bean;
-import io.inverno.app.ticket.internal.exception.PlanAlreadyExistsException;
 import io.inverno.app.ticket.internal.model.Plan;
 import io.inverno.app.ticket.internal.model.Ticket;
 import io.inverno.app.ticket.internal.rest.DtoMapper;
 import io.inverno.app.ticket.internal.rest.v1.dto.PlanDto;
 import io.inverno.app.ticket.internal.service.PlanService;
 import io.inverno.mod.base.resource.MediaTypes;
+import io.inverno.mod.http.base.ForbiddenException;
 import io.inverno.mod.http.base.Method;
 import io.inverno.mod.http.base.NotFoundException;
 import io.inverno.mod.http.base.Status;
 import io.inverno.mod.http.base.header.Headers;
+import io.inverno.mod.security.accesscontrol.PermissionBasedAccessController;
+import io.inverno.mod.security.http.context.SecurityContext;
+import io.inverno.mod.security.identity.Identity;
 import io.inverno.mod.web.base.annotation.Body;
 import io.inverno.mod.web.base.annotation.FormParam;
 import io.inverno.mod.web.base.annotation.PathParam;
@@ -150,10 +153,18 @@ public class PlanWebController {
 	 * @return
 	 */
 	@WebRoute( path = "/{planId}/ticket", method = Method.POST, consumes= MediaTypes.APPLICATION_X_WWW_FORM_URLENCODED )
-	public Mono<Void> pushTicket(@PathParam long planId, @FormParam long ticketId, @FormParam Optional<Long> referenceTicketId) {
-		return referenceTicketId
-			.map(refTicketId -> this.planService.insertTicketBefore(planId, ticketId, refTicketId))
-			.orElse(this.planService.addTicket(planId, ticketId));
+	public Mono<Void> pushTicket(@PathParam long planId, @FormParam long ticketId, @FormParam Optional<Long> referenceTicketId, SecurityContext<? extends Identity, ? extends PermissionBasedAccessController> securityContext) {
+		return securityContext.getAccessController()
+			.orElseThrow(() -> new ForbiddenException("Missing access controller"))
+			.hasPermission("push", "plan", Long.toString(planId))
+			.flatMap(hasPermission -> {
+				if(!hasPermission) {
+					throw new ForbiddenException();
+				}
+				return referenceTicketId
+					.map(refTicketId -> this.planService.insertTicketBefore(planId, ticketId, refTicketId))
+					.orElse(this.planService.addTicket(planId, ticketId));
+			});
 	}
 
 	/**
@@ -165,7 +176,15 @@ public class PlanWebController {
 	 * @return 1 if the ticket was removed, 0 if the ticket wasn't associated to the plan
 	 */
 	@WebRoute( path = "/{planId}/ticket/{ticketId}", method = Method.DELETE, produces = MediaTypes.TEXT_PLAIN )
-	public Mono<Long> removeTicket(@PathParam long planId, @PathParam long ticketId) {
-		return this.planService.removeTicket(planId, ticketId);
+	public Mono<Long> removeTicket(@PathParam long planId, @PathParam long ticketId, SecurityContext<? extends Identity, ? extends PermissionBasedAccessController> securityContext) {
+		return securityContext.getAccessController()
+			.orElseThrow(() -> new ForbiddenException("Missing access controller"))
+			.hasPermission("remove", "plan", Long.toString(planId))
+			.flatMap(hasPermission -> {
+				if(!hasPermission) {
+					throw new ForbiddenException();
+				}
+				return this.planService.removeTicket(planId, ticketId);
+			});
 	}
 }
